+++ src/wrap/append.js
@@ -1,1 +1,1 @@
-})(window);
+})(window.jQuery);