+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 #Changelog
 
+**master**
+
+  - fixed $ conflict (#525)
+
 **2.0.0-rc2**
 
   - added `data-parsley-validate-if-empty` field option (#489)